[#55] extract StudentsCrmV2.Concerns.Pager
