+++ backend/apps/students_crm_v2/lib/students_crm_v2/interactions/user/list.ex
@@ -1,6 +1,7 @@
 defmodule StudentsCrmV2.Interactions.User.List do
   @moduledoc false
 
+  alias StudentsCrmV2.Concerns.Pager
   alias StudentsCrmV2.Repo
 
   alias StudentsCrmV2.Models.{
@@ -10,12 +11,10 @@ defmodule StudentsCrmV2.Interactions.User.List do
 
   @spec execute(params :: map()) :: {[User.t()] | no_return(), map()}
   def execute(params) do
-    page_number = page_param(params, "number", "1")
-    page_size = page_param(params, "size", "50")
     tenant_id = Map.get(params, "tenant_id")
     query = by_tenant(User, tenant_id)
 
-    {page(query, page_number, page_size), %{total: count(query)}}
+    Pager.page(query, params)
   end
 
   defp by_tenant(query, tenant_id) do
@@ -26,19 +25,4 @@ defmodule StudentsCrmV2.Interactions.User.List do
     |> where([u, tu], tu.tenant_id == ^tenant_id)
     |> order_by(asc: :complex_name)
   end
-
-  defp page(query, page_number, page_size) do
-    import Ecto.Query, only: [limit: 2, offset: 2]
-
-    offset = page_size * (page_number - 1)
-
-    query
-    |> limit(^page_size)
-    |> offset(^offset)
-    |> Repo.all()
-  end
-
-  defp page_param(params, name, default), do: String.to_integer(get_in(params, ["page", name]) || default)
-
-  defp count(query), do: Repo.aggregate(query, :count, :id) || 0
 end
